slackbot/requirements.txt
# Core Slack and Databricks dependencies
slack-bolt>=1.23.0
slack-sdk>=3.36.0
databricks-sdk>=0.61.0

# Web framework and WSGI server
flask>=3.1.1
gunicorn>=23.0.0

# Additional dependencies for enhanced functionality
python-dotenv>=1.1.1

# Performance and monitoring
psutil>=7.0.0 